--- akka-docs/rst/java/code/docs/camel/sample/quartz/QuartzSample.java
@@ -1,12 +1,0 @@
-package docs.camel.sample.quartz;
-//#QuartzExample
-import akka.actor.ActorSystem;
-import akka.actor.Props;
-
-public class QuartzSample {
-  public static void main(String[] args) {
-    ActorSystem system = ActorSystem.create("my-quartz-system");
-    system.actorOf(Props.create(MyQuartzActor.class));
-  }
-}
-//#QuartzExample
