--- akka-docs/rst/java/code/docs/camel/sample/route/Producer1.java
@@ -1,10 +1,0 @@
-package docs.camel.sample.route;
-//#CustomRouteExample
-import akka.camel.javaapi.UntypedProducerActor;
-
-public class Producer1 extends UntypedProducerActor{
-  public String getEndpointUri() {
-    return "direct:welcome";
-  }
-}
-//#CustomRouteExample
