+++ akka-samples/akka-sample-camel-java/activator.properties
@@ -1,0 +1,4 @@
+name=akka-sample-camel-java
+title=Akka Camel Samples with Java
+description=Akka Camel Samples with Java
+tags=akka,camel,java,sample
