+++ akka-samples/akka-sample-camel-java/build.sbt
@@ -1,0 +1,14 @@
+name := "akka-sample-camel-java"
+
+version := "1.0"
+
+scalaVersion := "2.10.3"
+
+libraryDependencies ++= Seq(
+  "com.typesafe.akka" %% "akka-camel" % "2.3-SNAPSHOT",
+  "org.apache.camel" % "camel-jetty" % "2.10.3",
+  "org.apache.camel" % "camel-quartz" % "2.10.3",
+  "org.slf4j" % "slf4j-api" % "1.7.2",
+  "ch.qos.logback" % "logback-classic" % "1.0.7"
+)
+
