+++ akka-samples/akka-sample-camel-java/src/main/java/sample/camel/quartz/QuartzSample.java
@@ -1,0 +1,11 @@
+package sample.camel.quartz;
+
+import akka.actor.ActorSystem;
+import akka.actor.Props;
+
+public class QuartzSample {
+  public static void main(String[] args) {
+    ActorSystem system = ActorSystem.create("my-quartz-system");
+    system.actorOf(Props.create(MyQuartzActor.class));
+  }
+}
