+++ akka-samples/akka-sample-camel-java/src/main/java/sample/camel/route/RouteProducer.java
@@ -1,0 +1,9 @@
+package sample.camel.route;
+
+import akka.camel.javaapi.UntypedProducerActor;
+
+public class RouteProducer extends UntypedProducerActor {
+  public String getEndpointUri() {
+    return "direct:welcome";
+  }
+}
