+++ akka-samples/akka-sample-camel-java/tutorial/camel-async-interact.png
