+++ akka-samples/akka-sample-camel-java/tutorial/camel-custom-route.png
