+++ akka-samples/akka-sample-camel-scala/activator.properties
@@ -1,0 +1,4 @@
+name=akka-sample-camel-scala
+title=Akka Camel Samples with Scala
+description=Akka Camel Samples with Scala
+tags=akka,camel,scala,sample
