--- akka-samples/akka-sample-camel/README.md
@@ -1,15 +1,0 @@
-Camel Sample
-============
-
-This sample is meant to be used by studying the code; it does not perform any
-astounding functions when running it. If you want to run it, check out the akka
-sources on your local hard drive, follow the [instructions for setting up Akka
-with SBT](http://doc.akka.io/docs/akka/current/intro/getting-started.html).
-When you start SBT within the checked-out akka source directory, you can run
-this sample by typing
-
-    akka-sample-camel/run
-    
-and then choose which of the demonstrations you would like to run.
-
-You can read more in the [Akka docs](http://akka.io/docs).
