+++ akka-samples/akka-sample-main-java/activator.properties
@@ -1,4 +1,4 @@
 name=akka-sample-main-java
 title=Akka Main in Java
 description=Actor based version of obligatory Hello World program using the generic launcher class akka.Main. 
-tags=Basics,akka,java,java,starter
+tags=Basics,akka,java,starter
