+++ akka-samples/akka-sample-main-scala/activator.properties
@@ -1,4 +1,4 @@
 name=akka-sample-main-scala
 title=Akka Main in Scala
 description=Actor based version of obligatory Hello World program using the generic launcher class akka.Main. 
-tags=Basics,akka,java,scala,starter
+tags=Basics,akka,scala,starter
